+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - Support translation of optional types (`Optional<T>`, `OptionalInt`, `OptionalLong` and `OptionalDouble`).
   ([#15](https://github.com/EvidentSolutions/apina/issues/15))
 
+## Breaking changes
+
+- Change to Angular context: expose `apinaSerializationConfig` as `apinaConfig` instead.
+
 ## 0.2.2 (2015-06-15)
 
 ### Bug fixes